+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file.
 
+## 1.0.3
+### Renamed
+* database query to select
+### Added
+* database function for execute, executeReturn, and transaction
+* decimal type transformation when execute to database
+### Changed
+* remove onError and setError on every func on database, move it on result.
+
 ## 1.0.2
 ### Changed
 * do abstraction for all code